requirements.txt
fastapi
gradio
groq
pydantic
requests
uvicorn
python-dotenv